Fix a pile of font issues

+++ AppKit/NSFont.h
@@ -21,7 +21,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 @end
 
-typedef unsigned NSGlyph;
+typedef NSUInteger NSGlyph;
 
 enum {
     NSNullGlyph = 0,
